Address PR #864 round-4 review: kind in unique key, layer isolation (#869)
- Include `kind` in ReportTemplate uniqueness so a future JUDGE row can
  coexist with GENERATION at the same (variant, scope, communityId,
  version). Updates `@@unique` in schema.prisma, both unique indexes in
  the migration (full + partial SYSTEM-scope), and the DBML snapshot.
- Extract WeeklyReportPayload (and its nested AI-facing shapes) plus a
  new ReportVariant constant into src/application/domain/report/types.ts
  so the domain service no longer reaches into the presenter layer, and
  the weekly-report batch no longer imports GraphQL types for Prisma
  queries. service.ts / usecase.ts / presenter.ts and the two report
  unit-test files now import from the shared types module.
- Replace GqlReportStatus / GqlReportVariant in generateWeeklyReports.ts
  with Prisma's ReportStatus and the new ReportVariant constant; the
  batch is infrastructure code and must not depend on presentation types.
- Clarify SKIP_REASON_NO_ACTIVITY_PREFIX comment to spell out that the
  constant itself has no trailing colon (callers append ":" when
  building the persisted `skipReason` string).
