+++ src/application/domain/report/usecase.ts
@@ -3,7 +3,8 @@ import { inject, injectable } from "tsyringe";
 import { IContext } from "@/types/server";
 import { ValidationError } from "@/errors/graphql";
 import ReportService from "@/application/domain/report/service";
-import ReportPresenter, { WeeklyReportPayload } from "@/application/domain/report/presenter";
+import ReportPresenter from "@/application/domain/report/presenter";
+import { WeeklyReportPayload } from "@/application/domain/report/types";
 import { addDays, daysBetweenJst, truncateToJstDate } from "@/application/domain/report/util";
 import { renderPromptTemplate } from "@/application/domain/report/util/promptRenderer";
 import { LlmClient } from "@/infrastructure/libs/llm";
