+++ src/infrastructure/prisma/schema.prisma
@@ -1700,27 +1700,29 @@ model ReportTemplate {
   updatedAt DateTime? @updatedAt @map("updated_at")
 
   // Uniqueness / consistency invariants enforced at the DB layer:
-  //   1. @@unique([variant, communityId, version]) covers COMMUNITY-scope
-  //      rows (variant + non-NULL community_id + version triple). PostgreSQL
-  //      treats NULL as distinct in regular unique indexes, so this
-  //      attribute alone does NOT prevent duplicate SYSTEM-scope templates
-  //      on the same (variant, version).
+  //   1. @@unique([variant, communityId, kind, version]) covers COMMUNITY-scope
+  //      rows (variant + non-NULL community_id + kind + version quadruple).
+  //      `kind` is part of the key so a GENERATION row and a JUDGE row at the
+  //      same variant+version+community can coexist (LLM-as-Judge lands on
+  //      the same table in PR-F7+). PostgreSQL treats NULL as distinct in
+  //      regular unique indexes, so this attribute alone does NOT prevent
+  //      duplicate SYSTEM-scope templates on the same (variant, kind, version).
   //      (Ideally this would be `nullsNotDistinct: true` and absorb both
   //      scopes in a single constraint, but Prisma 6.11 only supports that
   //      argument on field-level `@unique`, not `@@unique`. The partial
   //      index below is the documented workaround until Prisma lands it.)
   //   2. The partial unique index `t_report_templates_variant_system_key`
   //      (defined in the corresponding migration as raw SQL
-  //      `CREATE UNIQUE INDEX ... ON (variant, version) WHERE community_id
-  //      IS NULL`) limits SYSTEM scope to one template per (variant, version)
-  //      pair.
+  //      `CREATE UNIQUE INDEX ... ON (variant, kind, version) WHERE
+  //      community_id IS NULL`) limits SYSTEM scope to one template per
+  //      (variant, kind, version) triple.
   //   3. The CHECK constraint `t_report_templates_scope_community_id_check`
   //      continues to enforce scope='SYSTEM' ⇔ community_id IS NULL and
   //      scope='COMMUNITY' ⇔ community_id IS NOT NULL, so (1) and (2) can
   //      rely on the scope/community_id pairing.
   //   4. `trafficWeight` is bounded 0..100 via
   //      `t_report_templates_traffic_weight_check` (see migration).
-  @@unique([variant, communityId, version])
+  @@unique([variant, communityId, kind, version])
   @@map("t_report_templates")
 }
